website-ui/src/app/create-job-page/execute/execute.component.ts
import { Component, OnInit } from '@angular/core';
import { MongodbService } from 'src/services/mongodb/mongodb.service';
import { Router, ActivatedRoute } from '@angular/router';
import { IJobModel } from '../../../../../mongodb-service/src/models/jobModel';

@Component({
  selector: 'app-execute',
  templateUrl: './execute.component.html',
  styleUrls: ['./execute.component.css']
})
export class ExecuteComponent implements OnInit {
  analysisJobIsExecuted = false;
  job: IJobModel;
  jobId: string;
  ioIsDisabled: boolean = true;

  constructor(public mongodbService: MongodbService, private route: ActivatedRoute, private router: Router) { }

  ngOnInit() {
    this.job = {
      name: "",
      description: "",
      rawInputDirectory: "",
      stagingFileName: "",
      userId: "",
      generateESIndices: true,
      jobStatus: 0,
    }

    this.route.params.subscribe(params => {
      this.jobId = params["job._id"];
      this.mongodbService.getJobById(this.jobId).subscribe(job => {
        this.job = job;
        if (this.job.jobStatus >= 5) {
          this.analysisJobIsExecuted = true;
        }
        this.ioIsDisabled = false;
      });
    });
  }

  submitAndRun() {
    this.mongodbService.updateJob(this.job).subscribe(retJob => {
      this.router.navigate(['/jobsPage']);
    });
  }

  deleteJob() {
    if (confirm("This job will be lost forever. Are you sure you want to delete it?")) {
      this.mongodbService.deleteJobRecusrive(this.job._id).subscribe(job => {
        this.router.navigate(['/jobsPage']);
      });
    }
  }
}
